Diferencia admin/cliente en vista principal y funciones de productos

+++ sof/cakephp/app/Controller/ProductsController.php
@@ -11,6 +11,12 @@ class ProductsController extends AppController
     public function index()
     {
         $this->set('products', $this->Product->find('all'));
+		  if($this->Session->read("Auth.User.role") == 'admin'){
+            $this->set('role','admin');
+        }
+        else{
+            $this->set('role','cust');
+        }
     }
 
     public function view($id = null)
@@ -66,41 +72,46 @@ class ProductsController extends AppController
     }*/
 	
 	public function edit($id = null) {
-        $this->set('platforms', $this->Platform->find('list'));
-        $this->set('categories', $this->Category->find('list'));
-        if (!$id) {
-            throw new NotFoundException(__('Invalid product'));
-        }
+		if($this->Session->read("Auth.User.role") == 'admin') {
+			$this->set('platforms', $this->Platform->find('list'));
+			$this->set('categories', $this->Category->find('list'));
+			if (!$id) {
+				throw new NotFoundException(__('Invalid product'));
+			}
 
-        $product = $this->Product->findById($id);
-        if (!$product) {
-            throw new NotFoundException(__('Invalid product'));
-        }
+			$product = $this->Product->findById($id);
+			if (!$product) {
+				throw new NotFoundException(__('Invalid product'));
+			}
 
-        if ($this->request->is(array('product', 'put'))) {
-            $this->Product->id = $id;
-            if ($this->Product->save($this->request->data)) {
-				$stock_id = $this->Stock->find('first', array('conditions' => array('Stock.product_id' == $id)));
-				$this->Stock->id = $stock_id; 
-				$this->Stock->saveField('amount', $this->request->data['Product']['amount']);
-                if($this->request->data['Product']['archivo']['error'] == 0 &&  $this->request->data['Product']['archivo']['size'] > 0){
-                    // Informacion del tipo de archivo subido $this->data['Product']['archivo']['type']
-                    //$destino = WWW_ROOT.'uploads'.DS;
-                    $destino = WWW_ROOT.'img'.DS;
-                    move_uploaded_file($this->request->data['Product']['archivo']['tmp_name'], $destino.$this->request->data['Product']['archivo']['name']);
-                    $id = $this->request->data['Product']['id'];
-                    $this->Product->read(null, $id);
-                    $this->Product->set('image', $this->request->data['Product']['archivo']['name']);
-                    $this->Product->save();
-                }
-                $this->Session->setFlash(__('El producto se ha actualizado.'));
-                return $this->redirect(array('action' => 'index'));
-            }
-            $this->Session->setFlash(__('No se pudo guardar los cambios.'));
-        }
+			if ($this->request->is(array('product', 'put'))) {
+				$this->Product->id = $id;
+				if ($this->Product->save($this->request->data)) {
+					$stock_id = $this->Stock->find('first', array('conditions' => array('Stock.product_id' == $id)));
+					$this->Stock->id = $stock_id; 
+					$this->Stock->saveField('amount', $this->request->data['Product']['amount']);
+					if($this->request->data['Product']['archivo']['error'] == 0 &&  $this->request->data['Product']['archivo']['size'] > 0){
+						// Informacion del tipo de archivo subido $this->data['Product']['archivo']['type']
+						//$destino = WWW_ROOT.'uploads'.DS;
+						$destino = WWW_ROOT.'img'.DS;
+						move_uploaded_file($this->request->data['Product']['archivo']['tmp_name'], $destino.$this->request->data['Product']['archivo']['name']);
+						$id = $this->request->data['Product']['id'];
+						$this->Product->read(null, $id);
+						$this->Product->set('image', $this->request->data['Product']['archivo']['name']);
+						$this->Product->save();
+					}
+					$this->Session->setFlash(__('El producto se ha actualizado.'));
+					return $this->redirect(array('action' => 'index'));
+				}
+				$this->Session->setFlash(__('No se pudo guardar los cambios.'));
+			}
 
-        if (!$this->request->data) {
-            $this->request->data = $product;
+			if (!$this->request->data) {
+				$this->request->data = $product;
+			}
+		}else{
+            $this->Session->setFlash(__('Acceso no permitido.'));
+            return $this->redirect(array('action' => 'index'));
         }
     }
 	
@@ -142,42 +153,52 @@ class ProductsController extends AppController
     //en amount viene la cantidad
     //en category viene el array de categorías
 	public function add() {
-		$this->set('platforms', $this->Platform->find('list'));
-        $this->set('categories', $this->Category->find('list'));
-        if ($this->request->is('post')) { 
-            $this->Product->create();
-            if ($this->Product->save($this->request->data)) {
-                $this->Product->Stock->save(['product_id'=>$this->Product->id, 'amount'=>$this->request->data['Product']['amount']]);
-				if($this->request->data['Product']['archivo']['error'] == 0 &&  $this->request->data['Product']['archivo']['size'] > 0){
-				  // Informacion del tipo de archivo subido $this->data['Product']['archivo']['type']
-				  //$destino = WWW_ROOT.'uploads'.DS;
-				  $destino = WWW_ROOT.'img'.DS;
-				  move_uploaded_file($this->request->data['Product']['archivo']['tmp_name'], $destino.$this->request->data['Product']['archivo']['name']);
-				  $id = $this->request->data['Product']['id'];
-				  $this->Product->read(null, $id);
-				  $this->Product->set('image', $this->request->data['Product']['archivo']['name']);
-				  $this->Product->save();
+		if($this->Session->read("Auth.User.role") == 'admin') {
+			$this->set('platforms', $this->Platform->find('list'));
+			$this->set('categories', $this->Category->find('list'));
+			if ($this->request->is('post')) { 
+				$this->Product->create();
+				if ($this->Product->save($this->request->data)) {
+					$this->Product->Stock->save(['product_id'=>$this->Product->id, 'amount'=>$this->request->data['Product']['amount']]);
+					if($this->request->data['Product']['archivo']['error'] == 0 &&  $this->request->data['Product']['archivo']['size'] > 0){
+					  // Informacion del tipo de archivo subido $this->data['Product']['archivo']['type']
+					  //$destino = WWW_ROOT.'uploads'.DS;
+					  $destino = WWW_ROOT.'img'.DS;
+					  move_uploaded_file($this->request->data['Product']['archivo']['tmp_name'], $destino.$this->request->data['Product']['archivo']['name']);
+					  $id = $this->request->data['Product']['id'];
+					  $this->Product->read(null, $id);
+					  $this->Product->set('image', $this->request->data['Product']['archivo']['name']);
+					  $this->Product->save();
 
+					}
+					$this->Session->setFlash(__('Your product has been saved.'));
+					return $this->redirect(array('action' => 'index'));
 				}
-                $this->Session->setFlash(__('Your product has been saved.'));
-                return $this->redirect(array('action' => 'index'));
-            }
-            $this->Session->setFlash(__('Unable to add your product.'));
+				$this->Session->setFlash(__('Unable to add your product.'));
+			}
+		}else{
+            $this->Session->setFlash(__('Acceso no permitido.'));
+            return $this->redirect(array('action' => 'index'));
         }
     }
 	
     public function delete($id)
     {
-        if ($this->request->is('get'))
-        {
-            throw new MethodNotAllowedException();
-        }
+		if($this->Session->read("Auth.User.role") == 'admin') {
+			if ($this->request->is('get'))
+			{
+				throw new MethodNotAllowedException();
+			}
 
-        if ($this->Product->delete($id))
-        {
-            /*$this->Session->setFlash(
-                __('The post with id: %s has been deleted.', h($id))
-            );*/
+			if ($this->Product->delete($id))
+			{
+				/*$this->Session->setFlash(
+					__('The post with id: %s has been deleted.', h($id))
+				);*/
+				return $this->redirect(array('action' => 'index'));
+			}
+		}else{
+            $this->Session->setFlash(__('Acceso no permitido.'));
             return $this->redirect(array('action' => 'index'));
         }
     }
